[Reports] Add method objects.Task.extend_results()
This patch introduces method Task.extend_results()
that will be in use in next patch.

This method processes task results and extends/reorganizes them:
  * adds `info' part for each scenario results, with
    aggregated data (like maximum/minimum duration, iterations
    count and so on)
  * iterations data is transformed into iterator

This data representation will be used by new reports charts
classes, since they use streaming data processing.

In other side, this method is a workaround, because aggregated
data and iterations iterator should be obtained as-is directly
from the database api, so this method is expected to be
simplified in future after planned database refactoring.

Change-Id: I8de5a6243f99ed5c7d9b4fa422afc5787e860387