Add resource fault detail in exception
This commit adds the fault attribute of resource to
exception message when the resource goes into error
state. This will allow easy lookup and tracking
of the actual cause of the error rather than manually
looking it up for each resource. Adding to exception
also allows the detailed error to be available in
the generated reports.

Change-Id: I1293bc5b2fd1c344f0752d5cb89ccbbcebb351fc