Add batching for samples creation in context
Performance Ceilometer scenarios required a great amount of data, that means
that if something happens during samples are storing the test fails.
Batching improves it. Storing with batching has the same rate as if whole
sample list is stored, but choosing batch size not greater then 1% of all
samples makes it losing not critical for the test.

Change-Id: I96cfbdb6f0f8c20363b4be11cac488f6459f81e2