[Reports] Reduce histogram charts weight to make report faster
This patch removes histogram view `One Half', which seems to be
rarely used but holds a lot of data for JavaScript.

For example, there is a listing of how much data is generated for
JavaScript by MainHistogram, on dummy scenario with 50000 iterations:

   Square Root Choice - 224 points
   Sturges Formula - 17 points
   Rice Rule - 74 points
   One Half - 25000 points!!!

Removing this view causes HTML report to work much more faster.
In some cases (like given above) this actually revives the report.

Change-Id: I6f51e72b93084f57eaf8cc5a107a0dd103824944