Creation a list of samples in sample generator
In previous version of sample generator samples were created one per
request. Because of ceilometer api has possibility to processing
requests with array of samples it will be good to use it to speed
up samples creation.

In this patch were added changes for sample generator. Now samples
that intended for one resource are placed to one request as list of
samples. It must speed up samples creation.

Co-Authored-By: Igor Degtiarov <idegtiarov@mirantis.com>
Change-Id: I1ae300aa1de2c41dde3890db5b0c98519a09eaf1