[Verify] `rally verify install/reinstall` with --no-tempest-venv
The `rally verify start` command can accept the --system-wide-install flag.
It is intended to provide a possibility to use the local environment instead
of Tempest virtual environment. The flag has a strange name, and it was decided
to use --no-tempest-venv for the `rally verify install/reinstall` commands
instead of --system-wide-install. The next patch will be dedicated to rename
--system-wide-install into --no-tempest-venv that is more logical.

In addition:
1. Some doc strings were changed to keep consistency between them;

2. Some doc strings were changed to allow users to better understand
for what options are.

3. The is_installed() method was changed to fix the following issue:
this method didn't work correctly when the --no-tempest-venv flag was passed
to `rally verify install` because it always checked existance of .venv directory.

Change-Id: I5541ac4b417152eb15b1f6277efa06c9879b4180