[Verify] Fixing issue with failures of Heat tests
If you take a look at one of latest full runs of Tempest tests, you will
see that about 15 Heat tests fail. It is because of misconfigured option
"instance_type" in the orchestration section and incorrect values of
"max_template_size" and "max_resources_per_stack" options in config.ini.
These options should have the same values like these ones have in the
heat-api.conf file. So it will be better to leave the default values for
them that are suitable for many deployments.

So now "instance_type" option is configured, and "max_template_size" and
"max_resources_per_stack" options were removed.

Change-Id: I0f46e9b365cfd59905986ceb0229e85984548b5e