[Verify] Fixing issue with create_isolated_networks = False
If you take a look at https://review.openstack.org/#/c/229623/ and the
corresponding test run [1], you will see that 70 tempest tests failed.
It is because the create_isolated_networks flag is set to False. The
wrong option value was introduced by the mentioned patch. We don't
need to set the value for the option because we have the corresponding
Tempest context that uses the following logic:

    if the cloud doesn't have any shared networks, the default behavior
    defined in Tempest will be used and Tempest itself will manage network
    resources, i.e Tempest will create networks, subnets, routers, etc;

    if the cloud has some shared networks, we will create our own shared
    network and specify its name in the Tempest config file to avoid the
    "Multiple possible networks found." error;

    if user configures the fixed_network_name option in the tempest.conf
    file manually, then we don't need to do anything.

So we should remove create_isolated_networks = False from the config.ini
file to fix the issue.

Also we should use use_dynamic_credentials = True instead of allow_tenant_
isolation = True because this option is deprecated [2].

[1] http://logs.openstack.org/23/229623/3/gate/gate-rally-dsvm-verify-full/3921e9c/rally-verify/7_verify_results_--html.html.gz
[2] https://github.com/openstack/tempest/blob/master/tempest/config.py#L63

Change-Id: I4fe74bf6b4f12fc09ba94f3bef68c304630a848f