[Verify] Fixing failures of test_server_basicops
This patch is intended to fix the following error:

testtools.matchers._impl.MismatchError: u'172.24.5.204' != "Please login as 'cirros' user, not as root\n\n"

In order to fix the test_server_basicops test we can add "cirros" to
the name of a cirros image that will be uploaded to the cloud. Or we
can configure the ssh_user_regex option [1], which is more complicated.
So this patch adds "cirros" to the name of the cirros image. Also this
patch explicitly sets the ssh_user_regex option to the default value
to allow users to configure this option when they use their own image.

[1] https://github.com/openstack/tempest/blob/master/tempest/config.py#L1215

Change-Id: I94abe3ae8738e322472dae78796ff36c4d6132ba