[Verify] Adding "xfail" mechanism for Tempest tests
This mechanism allows us to list some tests, that are expected to fail,
in a YAML file and these tests will have "xfail" status instead of "fail".
Tests that are in the YAML file and at the same time have "success" status
will be marked as "uxsuccess".

Test test_get_vnc_console[id-c6bc11bf-592e-4015-9319-1c98dc64daf5] was enabled
to test out "xfail" mechanism. This test fails because VNC console is disabled
in Devstack that is used for Tempest tests. So the mentioned test was added to
expected_failures.yaml.

Change-Id: If5a90ea6172f61eaadd640fea284aef4df7782bc