[Reports] Make HistogramChart.render() suitable for JS
Current HistogramChart.render() return value has legacy
structure and is very unfriendly for JavaScript.
However this return value is about JS processing only, so
it is changed by this patch to the structure which is
actually ready for rendering.

This change allows to simplify JS code and intriduce
AngularJS `widget' directive in next patch.

Change-Id: I54c7d5cfc133ed08582024f1c1bef07a2e9b95a2