Add note about the fact that Rally won't use testtools' assertions
Rally's core team thinks testtools is too big a dependency to be used by
Rally for performing assertions.  This commit results from a rejected
proposition to do so.  To avoid this to be wasted work, this commit
adds a note about this to discourage people from trying this again in the
future. See https://review.openstack.org/#/c/263759/ for details.

Change-Id: Ife05cb96b5bc92f4fc9b56d09bd16cc62fd10158