Integrate Rally & Alembic
Alembic allows us to change in future DB schema,
which we are going to do soon.

This patch adds under rally/common/db/sqlalchemy/migrations
required files for alembic as well as a first init migrations
that was autogenerated from models that we have.

As well this patch adds CLI commands for DB management

Change-Id: I7caa090aa5c4c6563b7e34d0d09baa039a3aa718
Co-Authored-By: Illia Khudoshyn <ikhudoshyn@mirantis.com>