[verify] fix rally task for verification
After we switched to new subunit parser[1], we missed to port tempest
scenarios to new interface.

Method parce_results[2] returns SubunitV2StreamResult instance instead
of tuple (<dict with total numbers>, <dict with tests>) if results are
presented and None instead of typle (None, None) if not.

This patch fixes checks and the way how results are proccesed in
tempest scenarios.

TODO: in future we need to add at least one tempest scenario to our
gates(I suppose it should be verify-full job).

[1] - Ifcfce827a0df1e75f58e198ee899d4bb85ca9376
[2] - https://github.com/openstack/rally/blob/0.1.2/rally/verification/tempest/tempest.py#L378

Change-Id: Iff8281cb42dba6be24a7e0a530d72de17a41807d