[Verify] Fixing 'test_server_basicops'
If we take a look, for example, at this test report [1], we will see
that the 'test_server_basicops' test fails. To fix the test we should
configure 'ssh_user_regex' option in tempest.conf. So this patch adds
two more regexes to 'ssh_user_regex' option to log into an instance
with correct username.

[1] http://logs.openstack.org/02/265902/2/check/gate-rally-dsvm-verify-full/78ebb06/rally-verify/7_verify_results_--html.html.gz

Change-Id: I15c8e5b0f86b566b75198b812e818db506e6afdf