Using 'dt' as alias for datetime imports
Many well-known libraries have a conventions regarding which
alias to use. Choosing a different one will not break the code
but makes it less readable and harder to maintain for other
developers. Unless there is specific reason, it is usually best
to use the recommended alias which most people are already
familiar with. For datetime this convention is dt.

Change-Id: Ifba7b7beafdac8377b98ce952b7d3b9dade01670
Closes-Bug: #1535786