[Verify] Don't create new image when image already exists
The current Tempest resources context creates an image in the cloud every
time when we run the tests and even when we run only one test. It is not
optimal. In order to avoid creating the image every time a new option for
`rally verify start` was added. So if we want to run tests, but we want
to use some existing image, we should execute the following command:

    $ rally verify start --discover

This option will tell Rally to search for the image with name that is
configured in the rally.conf file and the discovered image will be used
for the tests.

The next step will be to add the same functionality for flavors.

Change-Id: I0400d5451585ca3c1a82450a58b33125a2488b42