Make Rally cope with unversioned keystone URL
The way in which OSClient was constructing a keystone client was flawed
and using a deprecated method. It was also second guessing the
discovery process, which keystoneclient itself can handle on its own.

The main symptom of the flawed approach was that you were unable to use
keystone URLs in the catalog that didn't have the version appended,
e.g.:
  * http://foo.rally:5000/v3
  versus
  * http://foo.rally:5000/

The latter form is the correct form according to the Keystone
developers I spoke to.

With the change, the client version that's returned is now determined by
the keystoneclient library itself based on whether you supply a URL
with a version in it or not.

Change-Id: I3fc6f7b9724922ab7d47ce371578897da611b455