Make Rally cope with unversioned keystone URL
This is the second attempt to get this code merged after the previous
version was reverted apparently due to these bugs:
 * https://bugs.launchpad.net/rally/+bug/1545889
 * https://bugs.launchpad.net/rally/+bug/1547092
 * https://bugs.launchpad.net/rally/+bug/1544839

The only genuine bug was the first one - defining endpoints in your
deployment config broke things. The other two are not the fault of this
change as explained in the bugs' comments.

The new changes are to force a pre-auth using a different mechanism to
the previous approach that was calling the deprecated
client.authenticate() method. Instead, it's now calling
auth.get_access(session) which returns an auth_ref that we can store on
the client.

Note that this is all only needed because Rally needs the service_type,
which is only returned in the service_catalog and not the API's
enpoints.list().

Previous commit message below:
---
The way in which OSClient was constructing a keystone client was flawed
and using a deprecated method. It was also second guessing the
discovery process, which keystoneclient itself can handle on its own.

The main symptom of the flawed approach was that you were unable to use
keystone URLs in the catalog that didn't have the version appended,
e.g.:
  * http://foo.rally:5000/v3
  versus
  * http://foo.rally:5000/

The latter form is the correct form according to the Keystone
developers I spoke to.

With the change, the client version that's returned is now determined by
the keystoneclient library itself based on whether you supply a URL
with a version in it or not.

Closes-Bug: 1547143
Closes-Bug: 1545889

Change-Id: I31daf8c0bdb7d9329f14722fcd0f51cc23519158