Fixed imcompatible argument in service create with keystone v2/3 API
The keystone service create scenario will complain about wrong
argument type with keystone v3 API,
While the third argument is a boolean value indicating whether
the service is enabled in v3, it was a string value to describe
service description in v2.

This commit fixes the issue by simply using the keyword
for the description argument as it is accepted in both versions.

Change-Id: Iba1b5f09b8efbf6a73d34e913bc3083cf57d2451
Closes-Bug: #1469897