[Cliutils] Skip "n/a" value in cliutils.pretty_float_formatter
Currently pretty_float_formatter skips only value of None
(and replaces it with "n/a") however we can use this formatter
in some cases (will actually use in next patch) when we already
mark missed values with "n/a" so it is obvious that we have to
skip this value and not try round it like float.

Since formatter expect roundavle value only, then we simply return
"n/a" instead of any non-numeric value like None and "n/a".

Also this change rewrites formatter tests with DDT.

Change-Id: I6402df4654a5a33019fb9a92e2b80ea2f3cd7ff8