[Reports] Fix calc of running iterations in parallel
Load profile chart is calulated wrongly.
It shows more running interations in parallel than actually
are running.

We are spliting all duration to 100 segments, and we
are adding +1 to each segment that has intersection
with segment (iter_start, iter_end).

Basically if we have 3 iterations with next timestamps
(1, 2.5), (2, 3) (2.5, 4)
and chart of next segments [(1, 2), (2, 3), (3, 4)]
we will get result as [1, 3, 1] instead of [1,2,1]
which is expected

This patch changes algorithm of calculation of parallel
running iterations, which is way simpler.

If chart segment is fully inside iteration segment +1,
If chart segment is partially inside iteration segment +0.5
(we don't need to calulate actually part of chart segment
 in iteration segment, because of the nature of data)

Co-Authored-By: Boris Pavlovic <boris@pavlovic.me>
Co-Authored-By: Alexander Maretskiy <amaretskiy@mirantis.com>

Change-Id: Ie12b0d7293cf97b4bdc13aea436854475cbdba36