Fix filters in glance v1 wrapper
As in glance v2, filters are passed as a keyword argument. Unlike
glance v2, owner is a separate kwarg, not a generic filter. Our old
implementation accidentally worked because the only filter we ever
used was 'owner', but it would have broken as soon as we tried to pass
literally any other filter.

Change-Id: If190f6edaf36b64524c2b15fd9060a1dce0f98fc