Remove extraneous isfile check for task file
Originally this checked to ensure that the file existed and wasn't a
directory, but bc819cb changed to the more restrictive
os.path.isfile(), which broke the ability to use output
redirection. (E.g., 'rally task start --task <(cat *.yaml)') The whole
check is unnecessary, since open(task_file) returns perfectly useful
error messages itself, and works in all the cases we care about -- we
don't actually care if something is a file, only if we can read it.

Change-Id: I949f5642744fc16e1b4cf167feeb2ecb867b6ad9