[Verify] Add 'version' arg to install/reinstall cmd
For now Rally always installs Tempest from the master branch. It is not
very convenient because Tempest may be broken in the master branch and
it would be very useful to use some commit or tag for Tempest.
So this commit adds a possibility to checkout the specified commit or
tag before Tempest installation.

Change-Id: Icfbb4221005466139c14d3f4aae34bbb3e013982