[Verify] Correctly configure auth URLs and auth version
When we have an auth URL, for example, like this http://127.0.0.1:5000/,
Rally fails to correctly configure the following tempest options:
auth_version, uri and uri_v3 and we may encounter the following error:

    InvalidConfiguration: Invalid Configuration
    Details: Unsupported auth version:

because of not correctly configured options:

    auth_version =
    uri = http://127.0.0.1:5000/
    uri_v3 = http:/v3/v3127.0.0.1:5000/v3

It should be

    auth_version = v2
    uri = http://127.0.0.1:5000/v2.0
    uri_v3 = http://127.0.0.1:5000/v3

instead.

Tempest needs /v2.0 and /v3 at the end of URLs.

Change-Id: Ifddd4667f0cce06d07a8a002d4bb3bd97deaf200