[Magnum] Context and scenario for Kubernetes
Context and scenario to create pods and RCs

In this scenario, a Kubernetes cluster is created and
pods and replication controllers are launched.  These
will in turn launch docker instances.  The manifest
for the pods and replication controllers specifies
the docker image to be downloaded and used in the
containers.

The sample files will create nginx containers.

This scenario is intended to test the performance
of the Kubernetes as provisioned and configured by
Magnum.  There are many ways to configure the
cluster, therefore it would be helpful to detect
if any configuration can be tuned for better
performance.

Partially-Implements: blueprint benchmark-scenarios-for-magnum
Co-Authored-By: Ton Ngo <ton@us.ibm.com>
Change-Id: I3284f44ecce1f6b30087ad380b72da9ac41f21ce