Refactor osclients
Now keystoneclient are separeted into 2 parts: working with resources
and authorization part. Secand part was moved into keystoneauth1 repo.

That patch is intended to drop depraceted usage of keystoneclient for
authorization in favor of keystoneauth1. Furthermore, in the next patch
we are going to use keystoneauth1 session for initialization of clients
where its possible.

Also :
  * Remove deprecated register decorator from osclient
  * Use _get_endpoint() in proper places
  * Refactor create_client in Keystone
  * Refactor unit tests in test_osclients

Change-Id: If67dffcd27f9bc3e7484d02eac62035c82415c3a