[requirements] Add helper for managing requirements
While we are not depend on OpenStack global-requirements, we need to be synced
with it to be sure that Rally compatible with all OpenStack related stuff.

This patch provides new helper script, which checks versions from g-r, correct
license and latest version from pypi.

The usage is simple:

   # sync rally requirements with g-r
   tox -e requirements

   # just format rally requirements
   tox -e requirements -- --format

    # sync rally requirements with g-r & add upper limits for all packages
   tox -e requirements -- --add-uppers

Also, this patch formats Rally requirements in unified form, puts right
licenses to requirements and adds rally licenses in correct way in setup.cfg

Change-Id: Ia50694f636d9f80f08d90cc8133ebac8bb3f8671