[Verify] Adding 'add-options' arg to `rally verify genconfig` cmd
Tempest supports a mechanism that allows us to plug in various tests as
a Tempest plugin. Each plugin needs to be configured in the proper way.
Rally cannot configure all plugins on the planet. We should add some
option that will allow us to add extra configuration for plugins to
Tempest config file.

For example, if we want to run Ceilometer tests, the workflow will be
as follows:

$ rally verify install
$ rally verify installplugin --source https://github.com/openstack/ceilometer.git
$ rally verify genconfig --add-options /path/to/config/file/for/ceilometer/tests
$ rally verify start --regex ceilometer

where config file for Ceilometer tests is something like this:

...
[service_available]

ceilometer = True
event_enabled = True
...

Change-Id: Ida02a36e2757a742ca91919cc756846f137a23d3