[Verify] Fixing issue with system-wide Tempest installation
The issue was noticed in a Rally docker container. In the docker container,
as we know, Rally is installed in the global system environment. When we try
to execute

 $ rally verify install --system-wide

Tempest installation will be broken on

 $ pip install --no-deps -e ./

because of permission issues. The command requires the 'sudo' prefix. So it
looks like the best solution for resolving the issue is that user will take
care of installing Tempest package like he does for Tempest requirements.

Change-Id: I3c606d1f659600f027ecaf6fe99eae26c329c9a3