[Reports] Hide menu list if there is no data at all
If there is no data passed to trends report for some reason
(for example if all given tasks are failed) then there is an
error message "No data" is shown (it is OK) but there is also
an empty menu list is visible which should be hidden in this
case.

This patch fixes this.

Change-Id: Ie1d82a1a41718a9c976d5a485fcdcac4f4a02d80