[Plugins] Add executable availability check to instance_test.sh
This is related VMTasks.*runcommand* workloads which can use
standard shell script "instance_test.sh".
In some rare cases (like [1]) it is possible that wide-used
executable is not available on the VM instance. The result is
confusing error message from shell interpreter which obstructs
understanding what is actually wrong with the scenario (nothing
is wrong actually with the scenario, but the instance).

This patch adds a small piece of code to script instance_test.sh
which checks availability of all used executable commands and
outputs clear error message if some command is missed.

[1] https://bugs.launchpad.net/rally/+bug/1613144

Change-Id: I63db699b053c16dce14df919f72579af2f66fa3e