Moving of EC2, Fuel and Glance scenarios into separate classes
Since current scenario implementation transforms method to class
at runtime, it is reasonable to have class-based scenario
implementation which will be much simpler to use behind the scenes.

This class should be based on Scenario and do not break compatibility.
The only required method is run() which is actually a body of scenario.

This patch contains changes for following groups
(scenarios and tests):
/ec2/
/fuel/
/glance/

Change-Id: Idcee33340141cfa0995503332403c63352bfc6e1