Moving of Ironic and Keystone scenarios into separate classes
Since current scenario implementation transforms method to class at runtime,
it is reasonable to have class-based scenario implementation which will be
much simpler to use behind the scenes.

This class should be based on Scenario and do not break compatibility.
The only required method is run() which is actually a body of scenario.

Change-Id: Ib4c2aa5651d8ac0750b937146e59376a7efcece0