Moving of Nova scenarios into separate classes
Since current scenario implementation transforms method to class
at runtime, it is reasonable to have class-based scenario
implementation which will be much simpler to use behind the scenes.

This class should be based on Scenario and do not break compatibility.
The only required method is run() which is actually a body of scenario.

Change-Id: If6d69aa6b16d1aa3745422e772bca98e6c4eaf9a