Add `extra` parameter to ExistingCloud deployment
Sometimes user needs to specify some extra options related to deployment.
Previously it was possible, but after we merged strict json validation,
users are not able to do it.

Let's provide the way to do it.

Also, this patch extends json schema with domain_name variable, since it is
used in several places.

Change-Id: I675086c97f2f34cf67b86576031f90bb13818ac3