Remove pycparser restriction from requirements
The assertion error mentioned to justify the restriction is
not actually due to this version of the python package, but
instead due to an improperly prepared wheel which was
uploaded to pypi. This wheel has now been removed and the
package works properly.

Reference: https://github.com/eliben/pycparser/issues/147

This patch removes the restriction so that Rally's
requirements are back in line with the OpenStack upper
constraint for pycparser.

Change-Id: I66ef46ad4c1511f39817270f59c7c1511645f2eb
Related-Bug: #1629830
Related-Bug: #1629726