Add missing device_owner for L3 HA's case
Patch Ieab53624dc34dc687a0e8eebd8477 added a new possible value for the
device_owner field of a port, that signifies a router's interface for a
subnet. The rally code was not changed accordingly, which causes rally
to try to delete the port using the wrong API ('neutron port-delete'
instead of 'neutron router-interface-delete'), causing cleanup errors
and leftover resources.

Closes-Bug: #1562878
Change-Id: I65933650d613e1800541dfdb33714a50f5c13db7