Add expected variable to heat.create_client
Closes-Bug: 1661643
This adds the instantiation of "interface" which is a variable expected
by keystoneauth1.adapter but is not provided by
rally.Heat.osclient.create_client. These commits provide the missing
variable and include test units for heatclient.

This issue became apparent after another bug was found (1655728). This
was an issue pertaining to neutron-client but it highlighted the
possibility that this could occur with the other service clients.

The limitations of this code are that it solves the problem ONLY for
heatclient. Somebody needs to go in and investigate whether this issue
is possible for all the other clients (minus neutronclient and
heatclient). This obviously costs time and effort.

Commit messages follow:

osclient: Add ToDo as reminder to revisit issue

The explicit creation of endpoint_type is a short-term solution.
When python-heatclient deprecates endpoint_type, this section of
code should be removed.

osclients: Pass interface variable to heat client

heatclient uses keystoneauth1.adapter which expects the interface
variable. This is the same variable of endpoint_type. The
short-term solution is to provide both endpoint_type (already
included in the previous commit) and interface. See the PR comment
for more proposed solutions.

osclients: Fix ToDo format

osclients: Include interface in ToDo comment

test_osclients: Unit tests for Heat.create_client

A unit test for heat client creation is included.
A unit test for heat client creation and ensuring the existence
of instantiated keyword arguments is also included.

Change-Id: I373658e27b357920b8ff6cc38f90459d9623fc9a