[docs] Fix _get_info for function based scenarios
'is_classbased' property of scenario is missed in case of function
based scenarios. It applies to function itself, but actual plugin in
case of rally engine terms is a FuncPlugin class. We need to apply
is_classbased flag there to have an equal interface for class based
and function based scenarios.

Change-Id: I8863944023bc14e8e3f1a11af7b8d297a86ec76d