Change volumes size or time of cind ci
create_and_extend_volume and create_and_get_volume scenarios are
often failed, because the status error.
Maybe, it is the reason that there is not enough disk space left,
we try it.
Move create_volume to the first scenario, and reduce the concurrency
and times to avoid image volumes which be created more times.

Change-Id: I9c398effa0bbfdc693ac1197c4dc671e1f5e778a