[Test] Fix randomly failing unit test
Because of BackupHelper.__del__ that uses os.path.exists,
there is a probability that __del__ will happen during running
of test_generate_random_path and it will grab a value
from mock.side_effect that as a result will lead to an error.

Solution:
Replace side_effect tuple with lambda-function that will
react to a specific value instead of hard coded return values.

Change-Id: I965041049079520a0aa354ddef32aeba047195ba