[cleanup] Fix neutron quotas resource
Class NeutronQuotas represents quotas resources of Neutron.
Patch I66601978a5ac0b0c22935893f2247a5da8752419 changed the behaviour of all
quotas managers, but it did not take into accaunt that class NeutronQuotas
has NeutronMixin as a parent class with all neutron specific methods.

Merge of QuotaMixin and NeutronMixin produced an issue when a keystone
project resource (a resource which is used for identifing quotas) is
proccessed like a neutron resource.
In this case exception is raisen due to difference in specification
(keystone project resource is a python object, neutron resource is a
python dictionary).

This patch removes NeutronMixin from parent classes of NeutronQuotas

Change-Id: I8e2fbdb974dc181fdc0e48d593bd75e3e929d238