Fixed project creation with Keystone v3
The current behavior for Rally users and Keystone v3 allows only to
create test projects for the users only in "default" domain.
Such behavior leads to authentication errors for the Rally users if the
admin specifies project_domain_name with other value rather than
"default". This fix resolve this bug.

Signe-ooff-by: Anton Kremenetsky <akremenetsky@dev.rtsoft.ru>

Change-Id: Ia2207f0b6a32da2ece6aaef64fd3227e16fb25ec
Closes-Bug: #1680837