Split validation at 3 layers
We can group validation by 3 types: syntax (check that config is valid
for plugin in terms of jsonschema, required arguments, etc),
platform (check that deployment has credentials for launching
the plugin on specific environment), semantic(all others which operates
users: for example validate that specific service is enabled).

Splitting validation by these groups is required to support "validators
for validators". I mean that if we check that syntax of config is valid
and that we have credentials for all required platforms, we able to
execute more complex validators and ensure that they will not failed on
"random" failures like "AttributeError", "KeyError" and etc.

Change-Id: I1ea38403af6cfbd6b7765ae3c0caca12759bc571