Modify log marker functions's usage in an exception
The log marker functions (_LI(), _LW(), _LE(), and _LC()) must only be 
used when the message is only sent directly to the log. Anytime that the
message will be passed outside of the current context (for example as 
part of an exception) the _() marker function must be used.
Ref: https://docs.openstack.org/developer/oslo.i18n/guidelines.html
For example:
msg = _LE('There was an error.')
LOG.exception(msg)
raise LocalExceptionClass(msg)
Instead, use this style:

msg = _('There was an error.')
raise LocalExceptionClass(msg)

Change-Id: I6af0c72a013c7c6bc1db111908291fe8e4f79a2a